Minor refactor
Created separated files: parsers.py and utils.py
